+++ twilio/rest/api/v2010/account/available_phone_number/local.py
@@ -45,8 +45,8 @@ class LocalList(ListResource):
                near_number=values.unset, near_lat_long=values.unset,
                distance=values.unset, in_postal_code=values.unset,
                in_region=values.unset, in_rate_center=values.unset,
-               in_lata=values.unset, in_locality=values.unset, limit=None,
-               page_size=None):
+               in_lata=values.unset, in_locality=values.unset,
+               fax_enabled=values.unset, limit=None, page_size=None):
         """
         Streams LocalInstance records from the API as a generator stream.
         This operation lazily loads records as efficiently as possible until the limit
@@ -70,6 +70,7 @@ class LocalList(ListResource):
         :param unicode in_rate_center: The in_rate_center
         :param unicode in_lata: The in_lata
         :param unicode in_locality: The in_locality
+        :param bool fax_enabled: The fax_enabled
         :param int limit: Upper limit for the number of records to return. stream()
                           guarantees to never return more than limit.  Default is no limit
         :param int page_size: Number of records to fetch per request, when not set will use
@@ -100,6 +101,7 @@ class LocalList(ListResource):
             in_rate_center=in_rate_center,
             in_lata=in_lata,
             in_locality=in_locality,
+            fax_enabled=fax_enabled,
             page_size=limits['page_size'],
         )
 
@@ -113,8 +115,8 @@ class LocalList(ListResource):
              near_number=values.unset, near_lat_long=values.unset,
              distance=values.unset, in_postal_code=values.unset,
              in_region=values.unset, in_rate_center=values.unset,
-             in_lata=values.unset, in_locality=values.unset, limit=None,
-             page_size=None):
+             in_lata=values.unset, in_locality=values.unset,
+             fax_enabled=values.unset, limit=None, page_size=None):
         """
         Lists LocalInstance records from the API as a list.
         Unlike stream(), this operation is eager and will load `limit` records into
@@ -137,6 +139,7 @@ class LocalList(ListResource):
         :param unicode in_rate_center: The in_rate_center
         :param unicode in_lata: The in_lata
         :param unicode in_locality: The in_locality
+        :param bool fax_enabled: The fax_enabled
         :param int limit: Upper limit for the number of records to return. list() guarantees
                           never to return more than limit.  Default is no limit
         :param int page_size: Number of records to fetch per request, when not set will use
@@ -165,6 +168,7 @@ class LocalList(ListResource):
             in_rate_center=in_rate_center,
             in_lata=in_lata,
             in_locality=in_locality,
+            fax_enabled=fax_enabled,
             limit=limit,
             page_size=page_size,
         ))
@@ -178,8 +182,8 @@ class LocalList(ListResource):
              distance=values.unset, in_postal_code=values.unset,
              in_region=values.unset, in_rate_center=values.unset,
              in_lata=values.unset, in_locality=values.unset,
-             page_token=values.unset, page_number=values.unset,
-             page_size=values.unset):
+             fax_enabled=values.unset, page_token=values.unset,
+             page_number=values.unset, page_size=values.unset):
         """
         Retrieve a single page of LocalInstance records from the API.
         Request is executed immediately
@@ -201,6 +205,7 @@ class LocalList(ListResource):
         :param unicode in_rate_center: The in_rate_center
         :param unicode in_lata: The in_lata
         :param unicode in_locality: The in_locality
+        :param bool fax_enabled: The fax_enabled
         :param str page_token: PageToken provided by the API
         :param int page_number: Page Number, this value is simply for client state
         :param int page_size: Number of records to return, defaults to 50
@@ -226,6 +231,7 @@ class LocalList(ListResource):
             'InRateCenter': in_rate_center,
             'InLata': in_lata,
             'InLocality': in_locality,
+            'FaxEnabled': fax_enabled,
             'PageToken': page_token,
             'Page': page_number,
             'PageSize': page_size,