+++ twilio/rest/api/v2010/account/incoming_phone_number/__init__.py
@@ -175,7 +175,8 @@ class IncomingPhoneNumberList(ListResource):
                voice_method=values.unset, voice_url=values.unset,
                emergency_status=values.unset, emergency_address_sid=values.unset,
                trunk_sid=values.unset, identity_sid=values.unset,
-               phone_number=values.unset, area_code=values.unset):
+               address_sid=values.unset, phone_number=values.unset,
+               area_code=values.unset):
         """
         Create a new IncomingPhoneNumberInstance
 
@@ -198,6 +199,7 @@ class IncomingPhoneNumberList(ListResource):
         :param unicode emergency_address_sid: The emergency_address_sid
         :param unicode trunk_sid: Unique string to identify the trunk
         :param unicode identity_sid: Unique string that identifies the identity associated with number
+        :param unicode address_sid: Unique string that identifies the address associated with number
         :param unicode phone_number: The phone number
         :param unicode area_code: The desired area code for the new number
 
@@ -226,6 +228,7 @@ class IncomingPhoneNumberList(ListResource):
             'EmergencyAddressSid': emergency_address_sid,
             'TrunkSid': trunk_sid,
             'IdentitySid': identity_sid,
+            'AddressSid': address_sid,
         })
 
         payload = self._version.create(
@@ -406,7 +409,7 @@ class IncomingPhoneNumberContext(InstanceContext):
                voice_method=values.unset, voice_url=values.unset,
                emergency_status=values.unset, emergency_address_sid=values.unset,
                trunk_sid=values.unset, voice_receive_mode=values.unset,
-               identity_sid=values.unset):
+               identity_sid=values.unset, address_sid=values.unset):
         """
         Update the IncomingPhoneNumberInstance
 
@@ -431,6 +434,7 @@ class IncomingPhoneNumberContext(InstanceContext):
         :param unicode trunk_sid: Unique string to identify the trunk
         :param IncomingPhoneNumberInstance.VoiceReceiveMode voice_receive_mode: The voice_receive_mode
         :param unicode identity_sid: Unique string that identifies the identity associated with number
+        :param unicode address_sid: Unique string that identifies the address associated with number
 
         :returns: Updated IncomingPhoneNumberInstance
         :rtype: twilio.rest.api.v2010.account.incoming_phone_number.IncomingPhoneNumberInstance
@@ -457,6 +461,7 @@ class IncomingPhoneNumberContext(InstanceContext):
             'TrunkSid': trunk_sid,
             'VoiceReceiveMode': voice_receive_mode,
             'IdentitySid': identity_sid,
+            'AddressSid': address_sid,
         })
 
         payload = self._version.update(
@@ -559,6 +564,7 @@ class IncomingPhoneNumberInstance(InstanceResource):
         # Marshaled Properties
         self._properties = {
             'account_sid': payload['account_sid'],
+            'address_sid': payload['address_sid'],
             'address_requirements': payload['address_requirements'],
             'api_version': payload['api_version'],
             'beta': payload['beta'],
@@ -620,6 +626,14 @@ class IncomingPhoneNumberInstance(InstanceResource):
         :rtype: unicode
         """
         return self._properties['account_sid']
+
+    @property
+    def address_sid(self):
+        """
+        :returns: Unique string that identifies the address associated with number
+        :rtype: unicode
+        """
+        return self._properties['address_sid']
 
     @property
     def address_requirements(self):
@@ -856,7 +870,7 @@ class IncomingPhoneNumberInstance(InstanceResource):
                voice_method=values.unset, voice_url=values.unset,
                emergency_status=values.unset, emergency_address_sid=values.unset,
                trunk_sid=values.unset, voice_receive_mode=values.unset,
-               identity_sid=values.unset):
+               identity_sid=values.unset, address_sid=values.unset):
         """
         Update the IncomingPhoneNumberInstance
 
@@ -881,6 +895,7 @@ class IncomingPhoneNumberInstance(InstanceResource):
         :param unicode trunk_sid: Unique string to identify the trunk
         :param IncomingPhoneNumberInstance.VoiceReceiveMode voice_receive_mode: The voice_receive_mode
         :param unicode identity_sid: Unique string that identifies the identity associated with number
+        :param unicode address_sid: Unique string that identifies the address associated with number
 
         :returns: Updated IncomingPhoneNumberInstance
         :rtype: twilio.rest.api.v2010.account.incoming_phone_number.IncomingPhoneNumberInstance
@@ -907,6 +922,7 @@ class IncomingPhoneNumberInstance(InstanceResource):
             trunk_sid=trunk_sid,
             voice_receive_mode=voice_receive_mode,
             identity_sid=identity_sid,
+            address_sid=address_sid,
         )
 
     def fetch(self):