+++ CHANGES.md
@@ -3,6 +3,13 @@ twilio-python Changelog
 
 Here you can see the full list of changes between each twilio-python release.
 
+[2017-10-20] Version 6.8.1
+---------------------------
+**Api**
+- Add `address_sid` param to IncomingPhoneNumbers create and update
+- Add 'fax_enabled' option for Phone Number Search
+
+
 [2017-10-13] Version 6.8.0
 ---------------------------
 **Api**